+++ example_test.go
@@ -7,6 +7,7 @@ import (
 )
 
 func ExampleSlugify() {
+	slugify.ToLower = true
 	s := "Nín hǎo. Wǒ shì zhōng guó rén This is a test ---"
 	fmt.Println(slugify.Slugify(s))
 	// Output: nin-hao-wo-shi-zhong-guo-ren-this-is-a-test